+++ src/pages/NotFoundPage.tsx
@@ -1,9 +1,7 @@
-
-
-export default function NotFound() {
+export default function NotFoundPage() {
   return (
     <div>
       <h1>404 Not Found</h1>
     </div>
-  )
+  );
 }
